backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]>=0.31.1
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.9
sqlalchemy==2.0.25
pydantic[email]>=2.11.7
yt-dlp>=2024.3.10
brotli>=1.1.0
requests>=2.28.0
